--- a/mlfinlab/util/__init__.py
+++ b/mlfinlab/util/__init__.py
@@ -3,8 +3,7 @@ Utility functions. In particular Chapter20 code on Multiprocessing and Vectoriza
 """
 
 from mlfinlab.util.fast_ewma import ewma
-from mlfinlab.util.multiprocess import expand_call, lin_parts, mp_pandas_obj, nested_parts, process_jobs_mp, process_jobs_, \
-    report_progress
+from mlfinlab.util.multiprocess import expand_call, lin_parts, mp_pandas_obj, nested_parts, process_jobs, process_jobs_, process_jobs_ray, report_progress
 from mlfinlab.util.volatility import get_daily_vol, get_garman_class_vol, get_yang_zhang_vol, get_parksinson_vol
 from mlfinlab.util.volume_classifier import get_bvc_buy_volume
 from mlfinlab.util.generate_dataset import get_classification_data