# Description

Add support for [Ray](https://github.com/ray-project/ray) in the mlfinalb `multiprocess` module to speed up multi threaded calculations.

When the presence of the [Ray](https://github.com/ray-project/ray)  module is detected, it will be used in place of python `multiprocessing`. Otherwise we fall back to the default `multiprocessing` module.

Running the unit tests vanilla does not show the potential of [Ray](https://github.com/ray-project/ray)  as opposed to `multiprocessing`, because the performance increase gets offset by the overhead of initializing Ray.

For the purpose of showcasing the potential performance increase I have slightly modified some tests to run 100 times in a loop in an attempt to simulate a higher and closer to real world usage workload.
These modified tests can be found in the following branch:
[ray-support-showcase](https://github.com/Gunther-Schulz/mlfinlab/tree/ray-support-showcase)

To install the branch with pip run:
`pip install git+https://github.com/Gunther-Schulz/mlfinlab@ray-support-showcase`

To run all affected tests in batch, run: `python -m unittest test_labels.py test_sample_weights.py test_sampling.py test_sb_bagging.py`


My results running an Intel(R) Xeon(R) CPU E3-1240 v3 @ 3.40GHz with 32Gb RAM:


| Test(s) | Without Ray | With Ray |
|-|-|-|
| `python -m unittest test_labels.py test_sample_weights.py test_sampling.py test_sb_bagging.py` |  __418s__ | __193s__ |
| `python -m unittest test_labels.py` |  __148s__ | __47s__ |
| `python -m unittest test_sample_weights.py` |  __144s__ | __47s__ |
| `python -m unittest test_sampling.py` |  __35s__ | __8s__ |
| `python -m unittest test_sb_bagging.py` |  __217s__ | __141s__ |

# Notes
I have left [Ray](https://github.com/ray-project/ray) as an optional dependency. Let me know if you prefer a hard dependency.

## Type of change

Please delete options that are not relevant.

- [x] New feature (non-breaking change which adds functionality)


# How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce. Please also list any relevant details for your test configuration

- [x] `python -m unittest test_labels.py`
- [x] `python -m unittest test_sample_weights.py`
- [x] `python -m unittest test_sampling.py `
- [x] `python -m unittest test_sb_bagging.py`

**Test Configuration**:
* Manjaro Linux
* VSCode


# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] New and existing unit tests pass locally with my changes
- [x] I have added documentation