@Gunther-Schulz 2020-05-05
I am not sure why these commits show up in this PR. Should I resubmit or is this how it should be?

@WanABRO
Update hrp.py
dbcf6b0
@WanABRO
Update test_hrp.py
2688041
@WanABRO
Update hierarchical_risk_parity.rst
b622d92
@WanABRO
Update test_hrp.py
b04337b
@proskurin
Merge branch 'develop' into HRP_custom
0f16868
@WanABRO
Add seriated correlations
0ad0a7b
@WanABRO
define corr matrix before if
46104c2