@rspadim 2020-05-05
this is one of best multiprocessing libs i used with python

there's some others wrapping some methods of pandas (pandarel) but if we build from scratch a new code i prefer the use of ray