@bobcolner 2020-07-29
Sounds good, thanks for the insight.

Is the priority queue public (can you share) I may be able to contribute.