@Gunther-Schulz 2020-07-30
> Hi @bobcolner,
> 
> I am very hesitant to add any new dependencies to the library. We already have too many.
> 
> I like this PR. However, there is another priority queue which has our focus. When things free up, we can revisit this.
> 
> Best
> Jacques

Hi, That is why is a completely optional dependency. The code checks, if the Ray module is installed. If not it falls back to python multiprocessing.